Allow ProjectClaims to support multiple resources
In discussing with other projects about typical usage
calculation, it became apparent that supporting multiple
resources in a single claim is an important requirement.

This commit modifies the ProjectClaim API in backwards incompatible
ways so that users can associate multiple resources to a single
claim. Since this library is still technically unreleased and less
than version 1.0, backwards incompatible changes are acceptable.

Change-Id: Ib0b1ec34b0e82e4439e53f1992c37d8b2f99fb04
Closes-Bug: 1790931