WIP: Two level limit enforcer
Building on the flat enforce, added POC for two lovel enforcement.

Clearly needs more testing to shake out issues, but it shows how it
could work in a way that when there is no hierarchy its almost as
efficient as the flat enforcer. We need extra class to keystone about
the project hierarchy, but otherwise no extra counting of resources.

Change-Id: I4d69d0982b33222977eb8c2b1a586f02a5e41cae